Create Black Jack Game
This is a mini black jack game where initially the prompt message will appear at the top of the screen asking for your first card and second card. While doing so, it will add up the values of those two cards and ask for a third card if less than 21, depending on your answer, it will either, add another card [this is done randomly] or not add anymore. Languages used were JAVASCRIPT HTML: CSS

Bugs
1) Notification doesn't pop up so fixing that
2) It doesnt do any further action after third card has been added

If there are any other bugs that you would like me to fix, please you may as well just change it :)